bot/requirements.txt
python-telegram-bot>=21.11,<23.0
